requirements.txt
Flask>=1.0.2
Flask-Cors>=3.0.10
Flask-SocketIO>=5.2.0 
pyserial>=3.5
python-engineio>=4.3.4
python-socketio>=5.7.1
PyYAML>=6.0
requests>=2.28.1


flask_restful